Refactor checkout demo a little bit
