+++ demo/modules/Checkout/components/Select/index.js
@@ -1,12 +1,14 @@
 import React from 'react';
+import CerebralForm from 'cerebral-module-forms/react/HOC';
 import cx from 'classnames';
 
-function PureInput(props) {
+function PureSelect(props) {
   const id = props.id || props.field.join('.');
   const {containerClassName,
-         inputClassName,
+         selectClassName,
          label,
          errorMessage,
+         options,
          isRequired,
          isTouched,
          isValid,
@@ -20,10 +22,16 @@ function PureInput(props) {
   return (
     <div className={cx(containerClassName, containerClasses)}>
       <label htmlFor={id}>{label}</label>
-      <input {...otherProps}
-             id={id}
-             name={props.name || id}
-             className={inputClassName} />
+      <select {...otherProps}
+              id={id}
+              name={props.name || id}
+              className={selectClassName}>
+        {options.map((option, index) =>
+          <option key={index} value={option.value || option.text}>
+            {option.text}
+          </option>
+        )}
+      </select>
       <aside className='pure-form-message'>
         {isTouched ? errorMessage : null}
       </aside>
@@ -31,9 +39,9 @@ function PureInput(props) {
   );
 }
 
-PureInput.defaultProps = {
+PureSelect.defaultProps = {
   containerClassName: 'pure-u-1',
-  inputClassName: 'pure-u-1'
+  selectClassName: 'pure-u-1'
 };
 
-export default PureInput;
+export default CerebralForm(PureSelect);
